requirements.txt
requests
beautifulsoup4
pandas
tqdm
groq
python-dotenv
langchain
langchain-community
openai